+++ Example/Example/Sources/Business Logic/Service/BookService.swift
@@ -18,9 +18,9 @@ protocol BookService {
 
 final class BookServiceImpl: BookService {
     
-    let apiClient: ConcurrencyClient
+    let apiClient: Client
     
-    init(apiClient: ConcurrencyClient) {
+    init(apiClient: Client) {
         self.apiClient = apiClient
     }
