+++ Example/Example/Sources/Business Logic/ServiceLayer.swift
@@ -15,7 +15,7 @@ final class ServiceLayer {
     
     static let shared = ServiceLayer()
     
-    private(set) lazy var apiClient: ConcurrencyClient = AlamofireClient(
+    private(set) lazy var apiClient: Client = AlamofireClient(
         baseURL: URL(string: "https://library.mock-object.redmadserver.com/api/v1/")!,
         configuration: .ephemeral,
         responseObserver: { [weak self] request, response, data, error in
